backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=46.0.6
sqlalchemy==2.0.30
python-multipart>=0.0.18
websockets==12.0
